Fix compile error when using VS2012 toolchain.
BUG=0

Review URL: https://chromiumcodereview.appspot.com/22861009

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@218349 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 576f181e13b2fc3bd810bc9d24d77f03615a5c73

+++ win8/metro_driver/file_picker_ash.cc
@@ -152,7 +152,7 @@ HRESULT OpenFilePickerSession::SinglePickerDone(SingleFileAsyncOp* async,
       LOG(ERROR) << "NULL IStorageItem";
     }
   } else {
-    LOG(ERROR) << "Unexpected async status " << status;
+    LOG(ERROR) << "Unexpected async status " << static_cast<int>(status);
   }
   app_view_->OnOpenFileCompleted(this, success_);
   return S_OK;
@@ -203,7 +203,7 @@ HRESULT OpenFilePickerSession::MultiPickerDone(MultiFileAsyncOp* async,
       LOG(ERROR) << "NULL StorageFileVectorCollection";
     }
   } else {
-    LOG(ERROR) << "Unexpected async status " << status;
+    LOG(ERROR) << "Unexpected async status " << static_cast<int>(status);
   }
   app_view_->OnOpenFileCompleted(this, success_);
   return S_OK;
@@ -543,7 +543,7 @@ HRESULT SaveFilePickerSession::FilePickerDone(SaveFileAsyncOp* async,
       LOG(ERROR) << "NULL IStorageItem";
     }
   } else {
-    LOG(ERROR) << "Unexpected async status " << status;
+    LOG(ERROR) << "Unexpected async status " << static_cast<int>(status);
   }
   app_view_->OnSaveFileCompleted(this, success_);
   return S_OK;
@@ -612,7 +612,7 @@ HRESULT FolderPickerSession::FolderPickerDone(FolderPickerAsyncOp* async,
       LOG(ERROR) << "NULL IStorageItem";
     }
   } else {
-    LOG(ERROR) << "Unexpected async status " << status;
+    LOG(ERROR) << "Unexpected async status " << static_cast<int>(status);
   }
   app_view_->OnFolderPickerCompleted(this, success_);
   return S_OK;
